+++ test/util/connectSpec.js
@@ -19,43 +19,23 @@ describe('util::connect', () => {
 
   it('connects store listeners to the dispatcher', () => {
     const callback = () => {};
-    const tokens = connect([
-      {
-        name: 'FOO',
-        params: [],
+    const tokens   = connect({
+      FOO: {
         function: callback,
         dependencies: ['ID_2']
       }
-    ]);
+    });
 
     expect(tokens['FOO']).to.equal('ID_1');
     expect(Dispatcher.store['ID_1'].dependencies[0]).to.equal('ID_2');
   });
 
-  it('disallows multiple listeners for one event', () => {
-    expect(() => {
-      connect([
-        {
-          name: 'FOO',
-          params: [],
-          function: () => {},
-          dependencies: []
-        },
-        {
-          name: 'FOO'
-        }
-      ])
-    }).to.throw('Cannot attach multiple listeners to event "FOO"!');
-  });
-
   it('autocompletes dependencies', () => {
-    const tokens = connect([
-      {
-        name: 'FOO',
-        params: [],
+    const tokens = connect({
+      FOO: {
         function: () => {}
       }
-    ]);
+    });
 
     expect(Dispatcher.store[tokens['FOO']].dependencies.length).to.equal(0);
   });